move version facet group into separate component
