+++ src/components/SearchResults/Facets/FacetGroup.js
@@ -1,12 +1,10 @@
-import React, { useCallback, useMemo, useState, useContext } from 'react';
+import React, { useCallback, useState } from 'react';
 import { css, cx } from '@leafygreen-ui/emotion';
 import Icon from '@leafygreen-ui/icon';
 import { palette } from '@leafygreen-ui/palette';
 import { Body } from '@leafygreen-ui/typography';
-import Select from '../../Select';
-import SearchContext from '../SearchContext';
-import { theme } from '../../../theme/docsTheme';
-import FacetValue, { initChecked } from './FacetValue';
+import FacetVersionGroup from './FacetVersionGroup';
+import FacetValue from './FacetValue';
 
 // Facet options with the following ids will truncate the initial number of items shown
 const TRUNCATE_OPTIONS = ['programming_language', 'sub_product'];
@@ -43,71 +41,6 @@ const showMoreStyle = css`
 const showMoreGlyphStyle = css`
   color: ${palette.gray.dark2};
 `;
-
-// TODO: move to new file
-
-const selectStyle = css`
-  font-size: ${theme.fontSize.small};
-  line-height: 20px;
-  margin-bottom: 8px;
-  padding-left: 22px;
-
-  button {
-    height: 22px;
-  }
-`;
-
-const FacetVersionGroup = ({ facetOption: { name, id, options } }) => {
-  // create select options from options
-  const selectOptions = useMemo(
-    () => options.map((option) => ({ text: option.name, value: option.id, key: option.key })),
-    [options]
-  );
-  const { handleFacetChange, searchParams } = useContext(SearchContext);
-
-  const [value, setValue] = useState(
-    () => selectOptions.find(({ key, value }) => initChecked(searchParams, key, value))?.value || selectOptions[0].value
-  );
-
-  const optionsById = useMemo(
-    () =>
-      options.reduce((map, option) => {
-        map[option.id] = option;
-        return map;
-      }, {}),
-    [options]
-  );
-
-  const onChange = useCallback(
-    (newOption) => {
-      const selectedOption = optionsById[newOption.value];
-      const oldSelection = optionsById[value];
-      const facetsToUpdate = [];
-
-      // update to unselect current selection
-      facetsToUpdate.push({
-        key: oldSelection.key,
-        id: oldSelection.id,
-        checked: false,
-      });
-
-      // update to select new selection
-      facetsToUpdate.push({
-        key: selectedOption.key,
-        id: selectedOption.id,
-        checked: true,
-      });
-
-      setValue(selectedOption.id);
-      handleFacetChange(facetsToUpdate);
-    },
-    [handleFacetChange, optionsById, value]
-  );
-
-  return <Select className={cx(selectStyle)} value={value} choices={selectOptions} onChange={onChange}></Select>;
-};
-
-// END TODO: move to new file
 
 // Representative of a "facet-option" from search server response
 const FacetGroup = ({ facetOption: { name, id, options }, isNested = false, numSelectedChildren }) => {
